moved forecasting heur to a separate file
